+++ kamikaze/coeur/util/utils.cc
@@ -31,5 +31,5 @@ double time_dt()
 	struct timeval now;
 	gettimeofday(&now, nullptr);
 
-	return now.tv_sec + now.tv_usec * 1e-6;
+	return static_cast<double>(now.tv_sec) + static_cast<double>(now.tv_usec) * 1e-6;
 }
